requirements.txt
agno
streamlit
python-dotenv
gspread